requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
httpx==0.27.2
google-cloud-storage==2.18.2
